playwright-renderer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
playwright==1.44.0
Pillow==10.2.0
PyYAML==6.0.1
minio
